requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
gunicorn==21.2.0
resend==0.8.0
python-dotenv==1.0.1
fastmcp==0.1.0
